fix loadScript + supprsession de l'import prism, remplacé par un chargement par loadJs
